refactor: mover lógica de validação para utils
